Split DateTimeAssertionSpecs
